requirements.txt
mysql-connector-python==9.5.0
customtkinter==5.2.2
python-dotenv==1.0.1
packaging
